# 3x performance, resolved memory allocation problem and correct handling of double-character codepoints