--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,6 @@ obj
 py-codes
 
 *.user
+.vs/
+*.ncrunchsolution
+*.ncrunchproject