--- a/Convert.bat
+++ b/Convert.bat
@@ -1,0 +1,2 @@
+@python py2cs.py
+@copy unidecoder-decodemap.txt assets /y