--- a/Unidecoder.Characters.cs
+++ b/Unidecoder.Characters.cs
@@ -30,7 +30,6 @@ namespace Unidecode.NET
     {
         private static readonly string[][] characters;
         private static readonly int MaxDecodedCharLength;
-        private static readonly int MaxStringLengthForStackAlloc;
 
         static Unidecoder()
         {
@@ -549,7 +548,6 @@ namespace Unidecode.NET
                         if (str.Length > MaxDecodedCharLength)
                             MaxDecodedCharLength = str.Length;
                 }
-                MaxStringLengthForStackAlloc = STACKALLOC_BUFFER_SIZE / MaxDecodedCharLength - 1;
             }
         }
     }