--- a/benchmark/Unidecode.NET.Benchmark/Program.cs
+++ b/benchmark/Unidecode.NET.Benchmark/Program.cs
@@ -1,0 +1,11 @@
+using BenchmarkDotNet.Running;
+
+namespace Unidecode.NET.Benchmark;
+
+internal class Program
+{
+  static void Main(string[] args)
+  {
+    var summary = BenchmarkRunner.Run<Benchmarks>();
+  }
+}