--- a/src/Unidecode.NET.csproj
+++ b/src/Unidecode.NET.csproj
@@ -20,4 +20,7 @@
     <NoWarn>1591</NoWarn>
     <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
   </PropertyGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="..\assets\unidecoder-decodemap.txt" Link="unidecoder-decodemap.txt" />
+  </ItemGroup>
 </Project>