--- a/src/Unidecoder.Characters.cs
+++ b/src/Unidecoder.Characters.cs
@@ -30,7 +30,6 @@ namespace Unidecode.NET
     {
         private static readonly string[][] characters;
         private static readonly int MaxDecodedCharLength;
-        private static readonly int MaxStringLengthForStackAlloc;
 
         static Unidecoder()
         {
@@ -549,8 +548,6 @@ namespace Unidecode.NET
                         if (str.Length > MaxDecodedCharLength)
                             MaxDecodedCharLength = str.Length;
                 }
-                MaxStringLengthForStackAlloc = STACKALLOC_BUFFER_SIZE / MaxDecodedCharLength - 1;
             }
         }
     }
-