… BenchmarkResults.txt
… assets/unidecoder-decodemap.txt
… benchmark/Unidecode.NET.Benchmark/Unidecode.NET.Benchmark.csproj
… test/CompleteUnidecoderTest.cs
… test/SlowUnidecoderTest.cs
… test/Unidecode.NET.Tests.csproj
… test/UnidecoderTest.cs