@csm101 2023-07-03
I went on coding and fixed issues #14 and #15 .
for #14 i scrapped the Unidecode.Character.cs source file and initialized the characters by reading data from a new resource file linked in the assembly
for #15, since proper handling of all codepoints has a certain performance impact, I introduce a UnidecodeAlgorithm enum that lets you choose between the "fast" implementation and the "Complete", but slower implementation.

To be precise now you can do these things:
```
  "xyz".Unidecode() // uses one of the two algorithms depending on the static property Unidecode.Algorithm
  "xyz".Unidecode(UnidecodeAlgorithm.Fast); // uses the exact algorithm specified in the argument
  
  Unidecode.Algorithm = UnidecodeAlgorithm.Complete; //sets the algoritm for the parameterless version to "complete"
```
I also added a test case for the example provided in issue #15 

these are the update benchmark results:
```
|                   Method |       Mean |     Error |    StdDev |   Gen0 | Allocated |
|------------------------- |-----------:|----------:|----------:|-------:|----------:|
|     FastUnidecodeRussian |  41.808 ns | 0.2353 ns | 0.2201 ns | 0.0038 |      64 B |
| CompleteUnidecodeRussian | 108.688 ns | 2.1999 ns | 2.5335 ns | 0.0148 |     248 B |
|       FastUnidecodeAscii |  13.860 ns | 0.1878 ns | 0.1756 ns |      - |         - |
|   CompleteUnidecodeAscii |  66.209 ns | 0.9575 ns | 0.8956 ns | 0.0019 |      32 B |
|     UnidecodeRussianChar |   2.400 ns | 0.0662 ns | 0.0619 ns |      - |         - |
|       UnidecodeAsciiChar |   1.338 ns | 0.0033 ns | 0.0031 ns |      - |         - |
|     UnidecodeRussianRune |   1.747 ns | 0.0134 ns | 0.0125 ns |      - |         - |
|       UnidecodeAsciiRune |   1.436 ns | 0.0168 ns | 0.0157 ns |      - |         - |
```